refactor: evidence make equivocation private (#15420)
