+++ tests/integration/evidence/keeper/infraction_test.go
@@ -10,6 +10,7 @@ import (
 	"cosmossdk.io/x/evidence/keeper"
 	"cosmossdk.io/x/evidence/testutil"
 	"cosmossdk.io/x/evidence/types"
+	abci "github.com/cometbft/cometbft/abci/types"
 	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	"gotest.tools/v3/assert"
 
@@ -111,13 +112,16 @@ func TestHandleDoubleSign(t *testing.T) {
 
 	// double sign less than max age
 	oldTokens := f.stakingKeeper.Validator(ctx, operatorAddr).GetTokens()
-	evidence := &types.Equivocation{
-		Height:           0,
-		Time:             time.Unix(0, 0),
-		Power:            power,
-		ConsensusAddress: sdk.ConsAddress(val.Address()).String(),
+	evidence := abci.RequestBeginBlock{
+		ByzantineValidators: []abci.Misbehavior{{
+			Validator: abci.Validator{Address: val.Address(), Power: power},
+			Type:      abci.MisbehaviorType_DUPLICATE_VOTE,
+			Time:      time.Unix(0, 0),
+			Height:    0,
+		}},
 	}
-	f.evidenceKeeper.HandleEquivocationEvidence(ctx, evidence)
+
+	f.evidenceKeeper.BeginBlocker(ctx, evidence)
 
 	// should be jailed and tombstoned
 	assert.Assert(t, f.stakingKeeper.Validator(ctx, operatorAddr).IsJailed())
@@ -128,7 +132,7 @@ func TestHandleDoubleSign(t *testing.T) {
 	assert.Assert(t, newTokens.LT(oldTokens))
 
 	// submit duplicate evidence
-	f.evidenceKeeper.HandleEquivocationEvidence(ctx, evidence)
+	f.evidenceKeeper.BeginBlocker(ctx, evidence)
 
 	// tokens should be the same (capped slash)
 	assert.Assert(t, f.stakingKeeper.Validator(ctx, operatorAddr).GetTokens().Equal(newTokens))
@@ -175,11 +179,13 @@ func TestHandleDoubleSign_TooOld(t *testing.T) {
 	)
 	assert.DeepEqual(t, amt, f.stakingKeeper.Validator(ctx, operatorAddr).GetBondedTokens())
 
-	evidence := &types.Equivocation{
-		Height:           0,
-		Time:             ctx.BlockTime(),
-		Power:            power,
-		ConsensusAddress: sdk.ConsAddress(val.Address()).String(),
+	evidence := abci.RequestBeginBlock{
+		ByzantineValidators: []abci.Misbehavior{{
+			Validator: abci.Validator{Address: val.Address(), Power: power},
+			Type:      abci.MisbehaviorType_DUPLICATE_VOTE,
+			Time:      ctx.BlockTime(),
+			Height:    0,
+		}},
 	}
 
 	cp := f.app.BaseApp.GetConsensusParams(ctx)
@@ -187,7 +193,8 @@ func TestHandleDoubleSign_TooOld(t *testing.T) {
 	ctx = ctx.WithConsensusParams(cp)
 	ctx = ctx.WithBlockTime(ctx.BlockTime().Add(cp.Evidence.MaxAgeDuration + 1))
 	ctx = ctx.WithBlockHeight(ctx.BlockHeight() + cp.Evidence.MaxAgeNumBlocks + 1)
-	f.evidenceKeeper.HandleEquivocationEvidence(ctx, evidence)
+
+	f.evidenceKeeper.BeginBlocker(ctx, evidence)
 
 	assert.Assert(t, f.stakingKeeper.Validator(ctx, operatorAddr).IsJailed() == false)
 	assert.Assert(t, f.slashingKeeper.IsTombstoned(ctx, sdk.ConsAddress(val.Address())) == false)
