+++ x/evidence/CHANGELOG.md
@@ -28,3 +28,4 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 * (x/evidence) [14724](https://github.com/cosmos/cosmos-sdk/pull/14724) The `x/evidence` module is extracted to have a separate go.mod file which allows it be a standalone module. 
+* (keeper) [#15420](https://github.com/cosmos/cosmos-sdk/pull/15420) Move `BeginBlocker` to the keeper folder & make HandleEquivocation private
