+++ x/evidence/module.go
@@ -164,7 +164,7 @@ func (AppModule) ConsensusVersion() uint64 { return 1 }
 
 // BeginBlock executes all ABCI BeginBlock logic respective to the evidence module.
 func (am AppModule) BeginBlock(ctx sdk.Context, req abci.RequestBeginBlock) {
-	BeginBlocker(ctx, req, am.keeper)
+	am.keeper.BeginBlocker(ctx, req)
 }
 
 // AppModuleSimulation functions
